(dbg) symbols not found in proto?

+++ recipes/libdydx_v4_proto-devel/helpers/build.ps1
@@ -47,12 +47,26 @@ if ($DLL) {
   $LIB = $DLL.FullName -replace "-\d+.dll", ".lib"
   $LIB = $LIB -replace "Library\\bin", "Library\\lib"
 
+  # Check if the symbol exists in the .dll
+  $dllExports = dumpbin /exports $DLL.FullName | Select-String -Pattern "cosmos::base::v1beta1"
+  if (-not $dllExports) {
+    Write-Output "Symbol 'cosmos::base::v1beta1' not found in $($DLL.FullName)"
+    exit 1
+  }
+
   dumpbin /exports $DLL.FullName | Select-String -Pattern "^[0-9A-F]+\s+[0-9A-F]+\s+.*$" | ForEach-Object { $_.ToString().Split(" ")[3] } | Out-File -FilePath $DEF
 
   if ($env:target_platform -eq "win-64") {
       lib /def:$DEF /out:$LIB /machine:x64
   } else {
       lib /def:$DEF /out:$LIB /machine:aarch64
+  }
+
+  # Check if the symbol exists in the .lib
+  $libSymbols = dumpbin /linkermember:1 $LIB | Select-String -Pattern "cosmos::base::v1beta1"
+  if (-not $libSymbols) {
+    Write-Output "Symbol 'cosmos::base::v1beta1' not found in $($LIB)"
+    exit 1
   }
 
   Remove-Item $DEF
